fix: split and validate immutable configuration
